split rules::either into three files
